Allow specifying root directory to start comparing from
This library was originally built mostly for web purposes,
where an absolute path would usually be the server's root
directory (e.g. /var/www).
There's some assumptions built in that have their basis
there, like how it'll append the paths to the current
working directory.
That makes it hard to use this library for cli based
transformations, though, where the root would usually be
'/'  instead of cwd...
There's now an optional 3rd argument that can be passed to
the constructor to inform it the root is something else
(e.g. '/')